+++ test/dpt_tests/dpt_time_test.py
@@ -14,13 +14,13 @@ class TestDPTTime:
     #
     def test_from_knx(self):
         """Test parsing of DPTTime object from binary values. Example 1."""
-        assert DPTTime().from_knx((0x4D, 0x17, 0x2A)) == time.strptime(
+        assert DPTTime.from_knx((0x4D, 0x17, 0x2A)) == time.strptime(
             "13 23 42 2", "%H %M %S %w"
         )
 
     def test_to_knx(self):
         """Testing KNX/Byte representation of DPTTime object."""
-        raw = DPTTime().to_knx(time.strptime("13 23 42 2", "%H %M %S %w"))
+        raw = DPTTime.to_knx(time.strptime("13 23 42 2", "%H %M %S %w"))
         assert raw == (0x4D, 0x17, 0x2A)
 
     #
@@ -28,12 +28,12 @@ class TestDPTTime:
     #
     def test_to_knx_max(self):
         """Testing KNX/Byte representation of DPTTime object. Maximum values."""
-        raw = DPTTime().to_knx(time.strptime("23 59 59 0", "%H %M %S %w"))
+        raw = DPTTime.to_knx(time.strptime("23 59 59 0", "%H %M %S %w"))
         assert raw == (0xF7, 0x3B, 0x3B)
 
     def test_from_knx_max(self):
         """Test parsing of DPTTime object from binary values. Example 2."""
-        assert DPTTime().from_knx((0xF7, 0x3B, 0x3B)) == time.strptime(
+        assert DPTTime.from_knx((0xF7, 0x3B, 0x3B)) == time.strptime(
             "23 59 59 0", "%H %M %S %w"
         )
 
@@ -42,37 +42,37 @@ class TestDPTTime:
     #
     def test_to_knx_min(self):
         """Testing KNX/Byte representation of DPTTime object. Minimum values."""
-        raw = DPTTime().to_knx(time.strptime("0 0 0", "%H %M %S"))
+        raw = DPTTime.to_knx(time.strptime("0 0 0", "%H %M %S"))
         assert raw == (0x0, 0x0, 0x0)
 
     def test_from_knx_min(self):
         """Test parsing of DPTTime object from binary values. Example 3."""
-        assert DPTTime().from_knx((0x0, 0x0, 0x0)) == time.strptime("0 0 0", "%H %M %S")
+        assert DPTTime.from_knx((0x0, 0x0, 0x0)) == time.strptime("0 0 0", "%H %M %S")
 
     #
     # TEST INITIALIZATION
     #
     def test_to_knx_default(self):
         """Testing default initialization of DPTTime object."""
-        assert DPTTime().to_knx(time.strptime("", "")) == (0x0, 0x0, 0x0)
+        assert DPTTime.to_knx(time.strptime("", "")) == (0x0, 0x0, 0x0)
 
     def test_from_knx_wrong_size(self):
         """Test parsing from DPTTime object from wrong binary values (wrong size)."""
         with pytest.raises(ConversionError):
-            DPTTime().from_knx((0xF8, 0x23))
+            DPTTime.from_knx((0xF8, 0x23))
 
     def test_from_knx_wrong_bytes(self):
         """Test parsing from DPTTime object from wrong binary values (wrong bytes)."""
         with pytest.raises(ConversionError):
             # this parameter exceeds limit
-            DPTTime().from_knx((0xF7, 0x3B, 0x3C))
+            DPTTime.from_knx((0xF7, 0x3B, 0x3C))
 
     def test_from_knx_wrong_type(self):
         """Test parsing from DPTTime object from wrong binary values (wrong type)."""
         with pytest.raises(ConversionError):
-            DPTTime().from_knx((0xF8, "0x23"))
+            DPTTime.from_knx((0xF8, "0x23"))
 
     def test_to_knx_wrong_parameter(self):
         """Test parsing from DPTTime object from wrong string value."""
         with pytest.raises(ConversionError):
-            DPTTime().to_knx("fnord")
+            DPTTime.to_knx("fnord")