+++ xknx/dpt/dpt_4byte_float.py
@@ -5,7 +5,9 @@ They correspond to the the following KDN DPT 14 class.
 """
 from __future__ import annotations
 
+from math import ceil, log10
 import struct
+from typing import cast
 
 from xknx.exceptions import ConversionError
 
@@ -35,9 +37,17 @@ class DPT4ByteFloat(DPTBase):
         """Parse/deserialize from KNX/IP raw data (big endian)."""
         cls.test_bytesarray(raw)
         try:
-            return struct.unpack(">f", bytes(raw))[0]  # type: ignore
+            raw_float = cast(float, struct.unpack(">f", bytes(raw))[0])
         except struct.error:
             raise ConversionError("Could not parse %s" % cls.__name__, raw=raw)
+        try:
+            # round to 7 digit precicion independent of exponent - same value as ETS 5.7 group monitor
+            return round(raw_float, 7 - ceil(log10(abs(raw_float))))
+        except (ValueError, OverflowError):
+            # account for 0 and special values
+            # ValueError: log10(0.0); ceil(float('nan'))
+            # OverflowError: ceil(float('inf'))
+            return raw_float
 
     @classmethod
     def to_knx(cls, value: float) -> tuple[int, ...]: